+++ internal/service/mocks/OrdersService.go
@@ -165,7 +165,7 @@ func (_m *OrdersService) UpdateOrdersCalculated(ctx context.Context, order model
 }
 
 // UpdateOrdersStatus provides a mock function with given fields: ctx, order, status, accrual
-func (_m *OrdersService) UpdateOrdersStatus(ctx context.Context, order models.Order, status models.OrdersStatus, accrual int) (*models.Order, error) {
+func (_m *OrdersService) UpdateOrdersStatus(ctx context.Context, order models.Order, status models.OrdersStatus, accrual float64) (*models.Order, error) {
 	ret := _m.Called(ctx, order, status, accrual)
 
 	if len(ret) == 0 {
@@ -174,10 +174,10 @@ func (_m *OrdersService) UpdateOrdersStatus(ctx context.Context, order models.Or
 
 	var r0 *models.Order
 	var r1 error
-	if rf, ok := ret.Get(0).(func(context.Context, models.Order, models.OrdersStatus, int) (*models.Order, error)); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, models.Order, models.OrdersStatus, float64) (*models.Order, error)); ok {
 		return rf(ctx, order, status, accrual)
 	}
-	if rf, ok := ret.Get(0).(func(context.Context, models.Order, models.OrdersStatus, int) *models.Order); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, models.Order, models.OrdersStatus, float64) *models.Order); ok {
 		r0 = rf(ctx, order, status, accrual)
 	} else {
 		if ret.Get(0) != nil {
@@ -185,7 +185,7 @@ func (_m *OrdersService) UpdateOrdersStatus(ctx context.Context, order models.Or
 		}
 	}
 
-	if rf, ok := ret.Get(1).(func(context.Context, models.Order, models.OrdersStatus, int) error); ok {
+	if rf, ok := ret.Get(1).(func(context.Context, models.Order, models.OrdersStatus, float64) error); ok {
 		r1 = rf(ctx, order, status, accrual)
 	} else {
 		r1 = ret.Error(1)
