+++ internal/service/orders_service.go
@@ -89,7 +89,7 @@ func (os ordersService) UpdateOrdersCalculated(ctx context.Context, order models
 	return os.orderRepository.UpdateOrdersCalculated(ctx, order, calculated)
 }
 
-func (os ordersService) UpdateOrdersStatus(ctx context.Context, order models.Order, status models.OrdersStatus, accrual int) (*models.Order, error) {
+func (os ordersService) UpdateOrdersStatus(ctx context.Context, order models.Order, status models.OrdersStatus, accrual float64) (*models.Order, error) {
 	return os.orderRepository.UpdateOrdersStatus(ctx, order, status, accrual)
 }
